Sending time and vehicle status
It was found that during testing the info was not completely getting
send correctly. In trying to make less typing for the socket class
something has broken and is causes the game's framerate to drop
dramatically. Need to have a look at converting another way or going to
back to 2 line steps.