All Telemetry Info Getting Sent
All of the car telemetry info is now getting sent to the server as
documented in rfactorStates.txt. Default port in 50000, will in the
future add a settings.txt so that the port number can be changed if the
user wishes to. Ran for 40 minutes with no frame rate issues.